Fix build warnings

+++ test/performance/shmem_perf_suite/shmem_bw_atomics_perf.c
@@ -42,7 +42,7 @@
 #define uni_bw(len, metric_info, snode, NAME, TYPE, op)               \
     do {                                                                       \
         double start = 0.0, end = 0.0;                                         \
-        int i = 0, j = 0, num_itr = metric_info->trials + metric_info->warmup; \
+        unsigned long int i = 0, j = 0, num_itr = metric_info->trials + metric_info->warmup; \
         int dest = partner_node(*metric_info);                                 \
         shmem_barrier_all();                                                   \
                                                                                \
